평균평점 소숫점 표기, rating > content 추가

+++ bitcTeam3prj/src/main/java/bitc/fullstack405/bitcteam3prj/controller/MovieBoardController.java
@@ -61,13 +61,14 @@ public class MovieBoardController {
                 sum += rating.getMovieRating();
             }
 
-            avg = sum / size;
+            avg = (float) sum/ size;
             mv.addObject("avg", avg);
         }
 
 
         mv.addObject("movie", entity);
         mv.addObject("ratingList", ratingList);
+//        mv.addObject("content")
 
         MovieEntity movieEntity = movieService.selectMovieById(id);
